+++ example/cli/src/app/app-routing.module.ts
@@ -8,6 +8,10 @@ import { FieldsComponent } from './fields/fields.component';
 import { FilterComponent } from './filter/filter.component';
 import { EmptyComponent } from './empty/empty.component';
 import { RtlTreeComponent } from './rtl/rtl-tree.component';
+import { AsyncTreeComponent } from './async/async.component';
+import { SaveRestoreComponent } from './save-restore/save-restore.component';
+import { CheckboxesComponent } from './checkboxes/checkboxes.component';
+import { DragComponent } from './drag/drag.component';
 
 const routes: Routes = [
   {
@@ -38,6 +42,22 @@ const routes: Routes = [
     path: 'rtl',
     component: RtlTreeComponent
   },
+  {
+    path: 'async',
+    component: AsyncTreeComponent
+  },
+  {
+    path: 'save-restore',
+    component: SaveRestoreComponent
+  },
+  {
+    path: 'checkboxes',
+    component: CheckboxesComponent
+  },
+  {
+    path: 'drag',
+    component: DragComponent
+  }
 ];
 
 @NgModule({