+++ lib/models/tree.model.ts
@@ -1,10 +1,10 @@
 import { Injectable, EventEmitter } from '@angular/core';
-import { observable, computed, action } from 'mobx';
+import { observable, computed, action, autorun } from 'mobx';
 import { TreeNode } from './tree-node.model';
 import { TreeOptions } from './tree-options.model';
 import { TreeVirtualScroll } from './tree-virtual-scroll.model';
-import { ITreeModel } from '../defs/api';
-import { TREE_EVENTS, newName } from '../constants/events';
+import { ITreeModel, IDType, IDTypeDictionary } from '../defs/api';
+import { TREE_EVENTS } from '../constants/events';
 
 import first from 'lodash-es/first';
 import last from 'lodash-es/last';
@@ -27,10 +27,10 @@ export class TreeModel implements ITreeModel {
   virtualScroll: TreeVirtualScroll;
 
   @observable roots: TreeNode[];
-  @observable expandedNodeIds: { [id: string]: boolean } = {};
-  @observable activeNodeIds: { [id: string]: boolean } = {};
-  @observable hiddenNodeIds: { [id: string]: boolean } = {};
-  @observable focusedNodeId: string = null;
+  @observable expandedNodeIds: IDTypeDictionary = {};
+  @observable activeNodeIds: IDTypeDictionary = {};
+  @observable hiddenNodeIds: IDTypeDictionary = {};
+  @observable focusedNodeId: IDType = null;
   @observable virtualRoot: TreeNode;
 
   private firstUpdate = true;
@@ -42,15 +42,7 @@ export class TreeModel implements ITreeModel {
   // events
   fireEvent(event) {
     event.treeModel = this;
-    const newEventName = newName(event.eventName);
-    const deprecatedEvent = Object.assign({}, event, {
-      deprecated: `This event is deprecated, please use ${newEventName} instead`
-    });
-    event.eventName = newEventName;
-
-    this.events[deprecatedEvent.eventName].emit(deprecatedEvent);
     this.events[event.eventName].emit(event);
-    this.events.onEvent.emit(deprecatedEvent);
     this.events.event.emit(event);
   }
 
@@ -202,7 +194,7 @@ export class TreeModel implements ITreeModel {
         this._calculateExpandedNodes();
       }
     } else {
-      this.fireEvent({ eventName: TREE_EVENTS.onUpdateData });
+      this.fireEvent({ eventName: TREE_EVENTS.updateData });
     }
   }
 
@@ -264,15 +256,15 @@ export class TreeModel implements ITreeModel {
 
     if (value) {
       node.focus();
-      this.fireEvent({ eventName: TREE_EVENTS.onActivate, node });
+      this.fireEvent({ eventName: TREE_EVENTS.activate, node });
     } else {
-      this.fireEvent({ eventName: TREE_EVENTS.onDeactivate, node });
+      this.fireEvent({ eventName: TREE_EVENTS.deactivate, node });
     }
   }
 
   @action setExpandedNode(node, value) {
     this.expandedNodeIds = Object.assign({}, this.expandedNodeIds, {[node.id]: value});
-    this.fireEvent({ eventName: TREE_EVENTS.onToggleExpanded, node, isExpanded: value });
+    this.fireEvent({ eventName: TREE_EVENTS.toggleExpanded, node, isExpanded: value });
   }
 
   @action expandAll() {
@@ -285,6 +277,12 @@ export class TreeModel implements ITreeModel {
 
   @action setIsHidden(node, value) {
     this.hiddenNodeIds = Object.assign({}, this.hiddenNodeIds, {[node.id]: value});
+  }
+
+  @action setHiddenNodeIds(nodeIds) {
+    this.hiddenNodeIds = nodeIds.reduce((id, hiddenNodeIds) => Object.assign(hiddenNodeIds, {
+      [id]: true
+    }), {});
   }
 
   performKeyAction(node, $event) {
@@ -321,12 +319,12 @@ export class TreeModel implements ITreeModel {
     const ids = {};
     this.roots.forEach((node) => this._filterNode(ids, node, filterFn, autoShow));
     this.hiddenNodeIds = ids;
-    this.fireEvent({ eventName: TREE_EVENTS.onChangeFilter });
+    this.fireEvent({ eventName: TREE_EVENTS.changeFilter });
   }
 
   @action clearFilter() {
     this.hiddenNodeIds = {};
-    this.fireEvent({ eventName: TREE_EVENTS.onChangeFilter });
+    this.fireEvent({ eventName: TREE_EVENTS.changeFilter });
   }
 
   @action moveNode(node, to) {
@@ -355,7 +353,54 @@ export class TreeModel implements ITreeModel {
       to.parent.treeModel.update();
     }
 
-    this.fireEvent({ eventName: TREE_EVENTS.onMoveNode, node: originalNode, to: { parent: to.parent.data, index: toIndex } });
+    this.fireEvent({ eventName: TREE_EVENTS.moveNode, node: originalNode, to: { parent: to.parent.data, index: toIndex } });
+  }
+
+  @action copyNode(node, to) {
+    const fromIndex = node.getIndexInParent();
+
+    if (!this._canMoveNode(node, fromIndex , to)) return;
+
+    // If node doesn't have children - create children array
+    if (!to.parent.getField('children')) {
+      to.parent.setField('children', []);
+    }
+    const toChildren = to.parent.getField('children');
+
+    const nodeCopy = this.options.getNodeClone(node);
+
+    toChildren.splice(to.index, 0, nodeCopy);
+
+    node.treeModel.update();
+    if (to.parent.treeModel !== node.treeModel) {
+      to.parent.treeModel.update();
+    }
+
+    this.fireEvent({ eventName: TREE_EVENTS.copyNode, node: nodeCopy, to: { parent: to.parent.data, index: to.index } });
+  }
+
+  getState() {
+    return {
+      expandedNodeIds: this.expandedNodeIds,
+      activeNodeIds: this.activeNodeIds,
+      hiddenNodeIds: this.hiddenNodeIds,
+      focusedNodeId: this.focusedNodeId
+    };
+  }
+
+  @action setState(state) {
+    if (!state) return;
+
+    Object.assign(this, {
+      expandedNodeIds: state.expandedNodeIds || {},
+      activeNodeIds: state.activeNodeIds || {},
+      hiddenNodeIds: state.hiddenNodeIds || {},
+      focusedNodeId: state.focusedNodeId
+    });
+  }
+
+  subscribeToState(fn) {
+    autorun(() => fn(this.getState()));
   }
 
   // private methods
@@ -409,7 +454,7 @@ export class TreeModel implements ITreeModel {
     this.activeNodes
       .filter((activeNode) => activeNode !== node)
       .forEach((activeNode) => {
-        this.fireEvent({ eventName: TREE_EVENTS.onDeactivate, node: activeNode });
+        this.fireEvent({ eventName: TREE_EVENTS.deactivate, node: activeNode });
       });
 
     if (value) {